+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 > Juknum
 
 - Added /autopush command, act the same way as GetResults.js, but it can be now executed manually
+- Added /profile command, used to translate Discord ID into username, usefull for the website Gallery.
+- Added /hotfix command, basically a /test command but with a better name.
 
 #### February 10th 2021
 > RobertR11